Make possible pip install for dependencies
The downloading of the virtualenv and of the dependencies in the
Makefile were suppressed and summarized in requirements.txt, as this
process was the cause of many bugs.
pip install -r requirements.txt or other python package installers can
therefore be run to install the python dependencies.

Remain to be done: Writing a README for the installation process,
creating a environment call in Makefile to still be able to install a
virtualenv, and writing setup.py to make the compilation of curp only
depend on pip if wanted.