Create script python function (curp and cal_tc)
Currently no curp function can be properly called through python.
For example if I want to run curp main module in python I would type:
>>> import curp.curp
>>> curp.curp.main()
The goal would be that it becomes possible to type:
>>> import curp
>>> curp.curp("eflow.cfg", verbose=True)
and so on for every scripts which can be used in terminal.
This passes through modifiying the scripts themselves (halfway done)
and modifying __init__.py (to be done).