Implement Topology object using parmed Structure
parmed permits to read topology files for CHARMM, Amber and
Gromacs forcefields, hence instead of creating specific parsers,
a Topology object was built on top of their Structure object.
The IDE of the former Topology class was judged obsolete and
replaced. It should now also be replaced in curp parts using
topology objects.

The making of non-bonded table is bugged.