Fixed bugs introduced by kinetic energy flux updates
I was not familiar with the self variable used in classes.
I defined variables in the __init__ method using the self variable
and by doing so I was able to use that same variable in other methods
under the same class when ever these methods were called later in the
program.

I also fixed some namespace issues I introduced.