Implement TwoBodyForce fitted to new Topology
TwoBodyForce was made more pythonic through properties.
The lib_amberbase forces variable names were replaced by their
equivalent in parmed, and num_torsion term in torsion module (now
dihedral) was simply deleted as it was previously simply a list
of one, and no equivalent could be found in parmed library.

As improper and dihedral modules were exactly similar, and as
impropers and dihedrals in parmed for amber are regrouped in the
dihedrals group, the content of improper module was cleared, so
that the charmm improper formula could be put instead.

CHARMM and Amber both have specificities which remain to be
implemented, both in Topology and TwoBodyForces. As said, CHARMM
has impropers interactions which should be implemented, but also
Urey-Bradley and CMAP. On the other hand Amber has coulomb 1-4
and lennard-jones 1-4 energy terms. The corresponding constants
can be found in the dihedrals ff_cst terms (see documentation of
Topology).

It may be possible to implement the computation of these in
TwoBodyForces object in an almost automatical manner, without
creating inherited TwoBodyForcesAmber and TwoBodyForcesCharmm
objects, by changing in TopologyAmber and TopologyCharmm the keys of
bonded_inter and the results of decomp_list and then building
a method applied on the list given. Hence TopologyAmber and
TopologyCharmm should be fairly easy to build (once again, a semi-
automatic manner could be thought of by giving a list of attributes
from which the bonded Interactions objects should be built from,
which would avoid unnecessary classes and lower code complexity.)