Debug writer for develop branch
The problem is with the line that goes something like "if a ==
'something_A' or 'something_B'" because the correct way to write it is
of course "if a == 'something_A' or a == 'something_B'".

In the first line, the first argument, a == 'something_A' could be true
or faulse, but the second argument, 'something_B' is always true because
a string returns true if the string is not empty.